Maye mech branch (#22)
* Push with loads of mech stuff

Lots of mech stuff
Most units have entries and profiles.
Most have the correct points (unless I made any mistakes)
Most should have the correct general constraits.
Weapons and special rules need filling out.
No root files. Not many categories added.

Also the 4 demon engines and the 2 special characters need filling out.

* Linked to rules which were missing, but now in the files

Removed the placeholders I had made for special rules to code units.
Relinked them to the rules which had been uploaded to master afterwards.

* 3 - Added more units and some profiles

Added the last 6 units and restrictions.
Added unique profiles as well for those units.

* Added in the Apex Detachments

Added Apex Detachments. But needs to have the restrictions enforced and also some Prime slots need adding to GST to support some of them.

* Special Rules and a bit more

Added the text for unique special rules in the mech lib.
Renamed the detachments to match the others in the main file.
Added in the Arcana Benefice to all the Archmagos, Magos and Techpriests including those on Abeyants.

Added ability to select our sub-faction for all units which require the ability to select them.

Added Flareshield, Rad Furnace, Machinator Array wargear descriptions

* Added all the weapon profiles for Mech unique weapons

As title. Also moved some of the SEs to the unit which required it if it was a unique item.

* Fix to PTS of Mech Melee Weapons list

* Trying to fix Hetedoxists

* Another push to try to fix Malefic

Giving up for a while if this doesn't work.

* Max 1 on each cyber powers

Max 1 on each cybertheury powers

* 1 more go at this

Trying to get NR to work with modifier for unit categories

* 1 more go again

* test again

* Changed a lot of "Parents" to "Self"

Fix to make the Sub-faction work (Tested) once. Hopefully works

* Arcana Benefice fix

* Added Officer of the Line (2) category to Magos and Magos on Abeyant

Added Officer of the Line (2) category to Magos and Magos on Abeyant

* Fixes for 4D6

Changes for:
Special rules comments
Flatten Power Weapon entry
Max 1s in Mech Wargear
Triaros and Kara 2 weapons
Max 1 on Thallax weapons per model and Praetorian restriction. Additionally this also was done for Ursarax
Krios HKMs fix min2 to be a max2

* Moving stuff around and adding Gambit and Cyber stuff

Moved some Mech Wargear to the Wargear lib as it will likley appear elsewhere such as Knights and Custodes.
Moved Corpasant Stave to Weapons lib.
Added Gambit and Cybertheurgic Rite profiles to GST as these are required for clarity.
Added the missing categories such as Unstoppable and Rapid and Vehicle to GST to add to Mech units which require such things for Prime bonus checks.

* Test for Thallax

* Fixed detachment selection dropdown

---------

Co-authored-by: The4D6