The final Chainsword!!!!
Scrapped the previous idea with the Default Chainsword category and went back to having a seperate default chainsword entry on units but now have two versions of the Sergeant Melee Weapons list - one with chainswords for use to swap out anything OTHER than a chainsword (bolt pistols, bolters, shotguns etc) and one WITHOUT chainswords to use to swap out for a chainsword - that way we don'#t have to have any convoluted conditions - there are mutually opposing conditions when you have a single list that can't be got around with logic.

I've tested this against all units that have 'may exchange chainsword for...' and they work. Also works with lightning claw swaps for those units that take a pair.

Fixes #934