Changed Model Type/Sub-Type constraints from Unit to Model (#1108)
Fixes #1105

In the scramble to fix the deletion of the Model Type/Sub-Type categories from non-model entries yesterday, the replacement constraints were put in as at least 1 of x per Unit. This works fine in most instances except where both a 'sergeant' model and the normal models can take the same weapons list and the 'sergeant' should get additional options but the mooks shouldn't. Chaning the constraints to "Model is instance of <Model Type/SubType category>" fixes the issue, so I've replaced them all again!